Refactored Table of Contents
