+++ next/src/components/common/TableOfContents/utils.ts
@@ -10,5 +10,6 @@ export const handleOnClick = ({ id, headerOffset }: Props) => {
   const elementPosition = element?.getBoundingClientRect().top || 0 // current offset regarding to the current window scroll
   const windowOffset = window.scrollY
   const offsetPosition = elementPosition + windowOffset - headerOffset
+
   window.scrollTo({ top: offsetPosition, behavior: 'smooth' })
 }
